move bytes/hex string stuff to utils for now
